Formulario Orden de Compra: Se agrega el campo Descuento para digitar este valor. Se implementa el botón Desbloquear para el campo Descuento y el botón Productos para editar directamente desde el formulario. En los valores totales se incluye es de los descuentos. Se agregó el campo de texto Términos de Garantía. Es posible redimensionar la ventana del formulario. Formulario Cotizaciones: Se modificó el campo Estampillas para ingresar un valor numérico en lugar de porcentaje. Se incluyó el campo de totalización de Estampillas. Es posible redimensionar la ventana del formulario. Reporte Ordenes de Compra: Se agregaron los campos para Descuentos, Términos de Garantía y la Nota: Se anexan fichas con especificaciones técnicas. Se agregaron disparadores para las tablas principales para gestionar un Log de Auditoría.

+++ DAL/DALCotizacionDetalle.cs
@@ -12,10 +12,10 @@ namespace DAL
         private static CotizacionDetalleTableAdapter adaptador = new CotizacionDetalleTableAdapter();
 
         public static void InsertCotizacionDetalle(int IdCotizacion, int NumeroItem, string DetalleNumeroItem, int IdProducto, 
-            string DescripcionProducto, decimal Precio, float Cantidad, float PorcentajeIVA, float PorcentajeImpoconsumo, float Estampillas)
+            string DescripcionProducto, decimal Precio, float Cantidad, float PorcentajeIVA, float PorcentajeImpoconsumo, decimal Estampillas)
         {
             adaptador.InsertCotizacionDetalle(IdCotizacion, NumeroItem, DetalleNumeroItem, IdProducto, DescripcionProducto, Precio, Cantidad, 
-                PorcentajeIVA, PorcentajeImpoconsumo, (decimal)Estampillas);
+                PorcentajeIVA, PorcentajeImpoconsumo, Estampillas);
         }
 
         public static bool ProductoTieneMovimientoByIdProducto(int IdProducto)
